+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Allowed for zero-dimension edge cases in `cov_to_chol()` and `chol_to_cov()`.
 
+* Added function `map_indices()` that maps indices from an input vector to corresponding sequences of grouped indices.
+
 # oeli 0.7.0
 
 * Improved documentation of `diff_cov()`, `undiff_cov()`, and `delta()`.