+++ NEWS.md
@@ -1,4 +1,4 @@
-# oeli development
+# oeli 0.7.1
 
 * Allowed for zero-dimension edge cases in `cov_to_chol()` and `chol_to_cov()`.